Upgrade client and provider implementation

+++ influxdbv2/resource_create_buckets.go
@@ -181,7 +181,7 @@ func getRetentionRules(input interface{}) domain.RetentionRules {
 	for _, retentionRule := range retentionRulesSet {
 		rr, ok := retentionRule.(map[string]interface{})
 		if ok {
-			each := domain.RetentionRule{EverySeconds: rr["every_seconds"].(int)}
+			each := domain.RetentionRule{EverySeconds: rr["every_seconds"].(int64)}
 			result = append(result, each)
 		}
 	}
